split client and server crates
