+++ Cargo.toml
@@ -1,0 +1,2 @@
+[workspace]
+members = ["client", "server"]
