+++ client/Cargo.toml
@@ -44,7 +44,6 @@ path = "src/lib.rs"
 backtrace = "^0.2.1"
 chrono = "^0.2.21"
 data-encoding = "^1.1.2"
-docopt = "^0.6.78"
 error-chain = "0.1.12"
 futures = "^0.1"
 lazy_static = "^0.2.1"
@@ -53,7 +52,8 @@ mio = "^0.5.1"
 openssl = "^0.8.3"
 rand = "^0.3"
 rustc-serialize = "^0.3.18"
-rusqlite = "^0.7.3"
 time = "^0.1"
 tokio-core = "^0.1"
-toml = "^0.1"
+
+[dev-dependencies]
+trust-dns-server = { version = "*", path = "../server" }
