+++ client/src/error/mod.rs
@@ -21,37 +21,27 @@ mod encode_error;
 mod client_error;
 mod lexer_error;
 mod parse_error;
-mod persistence_error;
-mod config_error;
 
 pub use self::decode_error::Error as DecodeError;
 pub use self::encode_error::Error as EncodeError;
 pub use self::client_error::Error as ClientError;
 pub use self::lexer_error::Error as LexerError;
 pub use self::parse_error::Error as ParseError;
-pub use self::persistence_error::Error as PersistenceError;
-pub use self::config_error::Error as ConfigError;
 
 pub use self::decode_error::ErrorKind as DecodeErrorKind;
 pub use self::encode_error::ErrorKind as EncodeErrorKind;
 pub use self::client_error::ErrorKind as ClientErrorKind;
 pub use self::lexer_error::ErrorKind as LexerErrorKind;
 pub use self::parse_error::ErrorKind as ParseErrorKind;
-pub use self::persistence_error::ErrorKind as PersistenceErrorKind;
-pub use self::config_error::ErrorKind as ConfigErrorKind;
 
 pub use self::decode_error::ChainErr as DecodeChainErr;
 pub use self::encode_error::ChainErr as EncodeChainErr;
 pub use self::client_error::ChainErr as ClientChainErr;
 pub use self::lexer_error::ChainErr as LexerChainErr;
 pub use self::parse_error::ChainErr as ParseChainErr;
-pub use self::persistence_error::ChainErr as PersistenceChainErr;
-pub use self::config_error::ChainErr as ConfigChainErr;
 
 pub type DecodeResult<T> = Result<T, DecodeError>;
 pub type EncodeResult = Result<(), EncodeError>;
 pub type ClientResult<T> = Result<T, ClientError>;
 pub type LexerResult<T> = Result<T, LexerError>;
 pub type ParseResult<T> = Result<T, ParseError>;
-pub type PersistenceResult<T> = Result<T, PersistenceError>;
-pub type ConfigResult<T> = Result<T, ConfigError>;
