+++ server/src/authority/authority.rs
@@ -14,53 +14,17 @@
  * limitations under the License.
  */
 use std::collections::BTreeMap;
-use std::cmp::Ordering;
 
 use chrono::UTC;
 
-use ::authority::{Journal, RRSet, UpdateResult, ZoneType};
+use trust_dns::op::{Message, UpdateMessage, ResponseCode, Query};
+use trust_dns::rr::{DNSClass, Name, RData, Record, RecordType, RrKey, RrSet};
+use trust_dns::rr::rdata::{NSEC, SIG};
+use trust_dns::rr::dnssec::Signer;
+
+use ::authority::{Journal, UpdateResult, ZoneType};
 use ::error::{PersistenceErrorKind, PersistenceResult};
-use ::op::{Message, UpdateMessage, ResponseCode, Query};
-use ::rr::{DNSClass, Name, RData, Record, RecordType};
-use ::rr::rdata::{NSEC, SIG};
-use ::rr::dnssec::Signer;
 
-/// Accessor key for RRSets in the Authority.
-#[derive(Eq, PartialEq, Debug, Hash, Clone)]
-pub struct RrKey { name: Name, record_type: RecordType }
-
-impl RrKey {
-  /// Creates a new key to access the Authority.
-  ///
-  /// # Arguments
-  ///
-  /// * `name` - domain name to lookup.
-  /// * `record_type` - the `RecordType` to lookup.
-  ///
-  /// # Return value
-  ///
-  /// A new key to access the Authorities.
-  pub fn new(name: &Name, record_type: RecordType) -> RrKey {
-    RrKey{ name: name.clone(), record_type: record_type }
-  }
-}
-
-impl PartialOrd for RrKey {
-  fn partial_cmp(&self, other: &RrKey) -> Option<Ordering> {
-    Some(self.cmp(other))
-  }
-}
-
-impl Ord for RrKey {
-  fn cmp(&self, other: &Self) -> Ordering {
-    let order = self.name.cmp(&other.name);
-    if order == Ordering::Equal {
-      self.record_type.cmp(&other.record_type)
-    } else {
-      order
-    }
-  }
-}
 
 /// Authority is responsible for storing the resource records for a particular zone.
 ///
@@ -70,7 +34,7 @@ pub struct Authority {
   origin: Name,
   class: DNSClass,
   journal: Option<Journal>,
-  records: BTreeMap<RrKey, RRSet>,
+  records: BTreeMap<RrKey, RrSet>,
   zone_type: ZoneType,
   allow_update: bool,
   // Private key mapped to the Record of the DNSKey
@@ -95,7 +59,7 @@ impl Authority {
   /// # Return value
   ///
   /// The new `Authority`.
-  pub fn new(origin: Name, records: BTreeMap<RrKey, RRSet>, zone_type: ZoneType, allow_update: bool) -> Authority {
+  pub fn new(origin: Name, records: BTreeMap<RrKey, RrSet>, zone_type: ZoneType, allow_update: bool) -> Authority {
     Authority{ origin: origin, class: DNSClass::IN,  journal: None, records: records, zone_type: zone_type,
       allow_update: allow_update, secure_keys: Vec::new() }
   }
@@ -178,25 +142,28 @@ impl Authority {
     self.journal.as_ref()
   }
 
-  #[cfg(test)]
+  /// Enables the zone for dynamic DNS updates
   pub fn set_allow_update(&mut self, allow_update: bool) {
     self.allow_update = allow_update;
   }
 
-  #[cfg(test)]
+  /// Retrieve the Signer, which contains the private keys, for this zone
   pub fn get_secure_keys(&self) -> &[Signer] {
     &self.secure_keys
   }
 
+  /// Get the origin of this zone, i.e. example.com is the origin for www.example.com
   pub fn get_origin(&self) -> &Name {
     &self.origin
   }
 
+  /// What type is this zone
   pub fn get_zone_type(&self) -> ZoneType {
     self.zone_type
   }
 
-  pub fn get_records(&self) -> &BTreeMap<RrKey, RRSet> {
+  /// Get all the
+  pub fn get_records(&self) -> &BTreeMap<RrKey, RrSet> {
     &self.records
   }
 
@@ -209,18 +176,18 @@ impl Authority {
     self.lookup(&self.origin, RecordType::SOA, false).first().map(|v| *v)
   }
 
-  /// Returns the SOA record
-  ///
-  ///
+  /// Returns the SOA record for the zone
   pub fn get_soa_secure(&self, is_secure: bool) -> Vec<&Record> {
     self.lookup(&self.origin, RecordType::SOA, is_secure)
   }
 
+  /// Returns the minimum ttl (as used in the SOA record)
   pub fn get_minimum_ttl(&self) -> u32 {
     self.get_soa().map_or(0, |soa| if let &RData::SOA(ref rdata) = soa.get_rdata() { rdata.get_minimum() } else { 0 })
   }
 
-  fn get_serial(&self) -> u32 {
+  /// get the current serial number for the zone.
+  pub fn get_serial(&self) -> u32 {
     let soa = if let Some(ref soa_record) = self.get_soa() {
       soa_record.clone()
     } else {
@@ -307,7 +274,7 @@ impl Authority {
   ///   NONE     rrset    empty    RRset does not exist
   ///   zone     rrset    rr       RRset exists (value dependent)
   /// ```
-  fn verify_prerequisites(&self, pre_requisites: &[Record]) -> UpdateResult<()> {
+  pub fn verify_prerequisites(&self, pre_requisites: &[Record]) -> UpdateResult<()> {
     //   3.2.5 - Pseudocode for Prerequisite Section Processing
     //
     //      for rr in prerequisites
@@ -443,7 +410,7 @@ impl Authority {
   ///   and restore the zone to its original state before answering the
   ///   requestor.
   /// ```
-  fn authorize(&self, update_message: &Message) -> UpdateResult<()> {
+  pub fn authorize(&self, update_message: &Message) -> UpdateResult<()> {
     // 3.3.3 - Pseudocode for Permission Checking
     //
     //      if (security policy exists)
@@ -527,7 +494,7 @@ impl Authority {
   ///   MAILB, or any other QUERY metatype besides ANY, or any unrecognized
   ///   type, else signal FORMERR to the requestor.
   /// ```
-  fn pre_scan(&self, records: &[Record]) -> UpdateResult<()> {
+  pub fn pre_scan(&self, records: &[Record]) -> UpdateResult<()> {
     // 3.4.1.3 - Pseudocode For Update Section Prescan
     //
     //      [rr] for rr in updates
@@ -603,7 +570,7 @@ impl Authority {
   /// * `records` - set of record instructions for update following above rules
   /// * `auto_sign` - if true, the zone will auto_sign (assuming there are signers present), this
   ///                 should be disabled during recovery.
-  fn update_records(&mut self, records: &[Record], auto_sign: bool) -> UpdateResult<bool> {
+  pub fn update_records(&mut self, records: &[Record], auto_sign: bool) -> UpdateResult<bool> {
     let mut updated = false;
     let serial: u32 = self.get_serial();
 
@@ -760,7 +727,7 @@ impl Authority {
     assert_eq!(self.class, record.get_dns_class());
 
     let rr_key = RrKey::new(record.get_name(), record.get_rr_type());
-    let records: &mut RRSet = self.records.entry(rr_key).or_insert(RRSet::new(record.get_name(), record.get_rr_type(), serial));
+    let records: &mut RrSet = self.records.entry(rr_key).or_insert(RrSet::new(record.get_name(), record.get_rr_type(), serial));
 
     records.insert(record, serial)
   }
@@ -1073,448 +1040,75 @@ impl Authority {
   }
 }
 
-#[cfg(test)]
-pub mod authority_tests {
-  use std::collections::BTreeMap;
-  use std::net::{Ipv4Addr,Ipv6Addr};
-
-  use ::authority::ZoneType;
-  use ::rr::*;
-  use ::rr::rdata::{ NULL, SOA, TXT };
-  use ::op::*;
-  use super::*;
-
-  pub fn create_example() -> Authority {
-    let origin: Name = Name::parse("example.com.", None,).unwrap();
-    let mut records: Authority = Authority::new(origin.clone(), BTreeMap::new(), ZoneType::Master, false);
-    // example.com.		3600	IN	SOA	sns.dns.icann.org. noc.dns.icann.org. 2015082403 7200 3600 1209600 3600
-    records.upsert(Record::new().name(origin.clone()).ttl(3600).rr_type(RecordType::SOA).dns_class(DNSClass::IN).rdata(RData::SOA(SOA::new(Name::parse("sns.dns.icann.org.", None).unwrap(), Name::parse("noc.dns.icann.org.", None).unwrap(), 2015082403, 7200, 3600, 1209600, 3600 ))).clone(), 0);
-
-    records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()) ).clone(), 0);
-    records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("b.iana-servers.net.", None).unwrap()) ).clone(), 0);
-
-    // example.com.		60	IN	TXT	"v=spf1 -all"
-    //records.upsert(origin.clone(), Record::new().name(origin.clone()).ttl(60).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT{ txt_data: vec!["v=spf1 -all".to_string()] }).clone());
-    // example.com.		60	IN	TXT	"$Id: example.com 4415 2015-08-24 20:12:23Z davids $"
-    records.upsert(Record::new().name(origin.clone()).ttl(60).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["$Id: example.com 4415 2015-08-24 20:12:23Z davids $".to_string()]))).clone(), 0);
-
-    // example.com.		86400	IN	A	93.184.216.34
-    records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone(), 0);
-
-    // example.com.		86400	IN	AAAA	2606:2800:220:1:248:1893:25c8:1946
-    records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(), 0);
-
-    // TODO support these later...
-
-    // example.com.		3600	IN	RRSIG	NSEC 8 2 3600 20150926015219 20150905040848 54108 example.com. d0AXd6QRITqLeiYbQUlJ5O0Og9tSjk7IlxQr9aJO+r+rc1g0dW9i9OCc XXQxdC1/zyubecjD6kSs3vwxzzEEupivaKHKtNPXdnDZ5UUiaIC1VU9l 9h/ik+AR4rCTY6dYPCI6lafD/TlqQLbpEnb34ywkRpl5G3pasPrwEY7b nrAndEY=
-    // example.com.		3600	IN	NSEC	www.example.com. A NS SOA TXT AAAA RRSIG NSEC DNSKEY
-    // example.com.		86400	IN	RRSIG	NS 8 2 86400 20150915033148 20150824191224 54108 example.com. O2TCB5/v/b1XGlTQEj0/oGKp7dTueQ7zRmCtADDEDWrzLdWrKcmDGF37 mgKejcAlSYVhWLxyLlet7KqJhLu+oQcDTNf/BT3vNX/Ivx3sKhUUMpfi 8Mn5zhRqM9gbzZVCS/toJIYqOBqvAkS7UpkmpLzl0Zt2h4j0Gp/8GwRb ZU67l6M=
-    // example.com.		86400	IN	RRSIG	AAAA 8 2 86400 20150914212400 20150824191224 54108 example.com. AHd2BDNjtg4jPRQwyT4FHtlVTZDZ6IIusYVGCzWfnt5SZOoizyXnJhqX 44MeVTqi1/2cskpKvRkK3bkYnVUcjZiFgSaa9xJHmXrslaTr5mOmXt9s 6k95N1daYKhDKKcr0M4TXLUgdnBr+/pMFiLsyOoDb8GJDT8Llmpk52Ie ysJX8BY=
-    // example.com.		86400	IN	RRSIG	A 8 2 86400 20150914083326 20150824191224 54108 example.com. La1p2R7GPMrXEm3kcznSJ70sOspmfSDsgOZ74GlzgaFfMRveA20IDUnZ /HI9M95/tBWbHdHBtm9aCK+4n7EluhNPTAT1+88V6xK7Lc7pcBfBXIHg DAdUoj26VIh7NRml/0QR0dFu4PriA/wLNe+d1Q961qf0JZP80TU4IMBC X/W6Ijk=
-    // example.com.		60	IN	RRSIG	TXT 8 2 60 20150914201612 20150824191224 54108 example.com. Be/bPvaVVK/o66QOHJZMFBDCQVhP44jptS9sZe8Vpfmzd72/v+1gwn1z u2+xisePSpAMtDZsFJgqsCjpbLFvmhNdh8ktlq/kuCME5hZs7qY7DZIB VwkSTsJPIq8qhX22clfIbqzaypuIX9ajWr+5i0nGQLNekMB07t4/GCoJ q5QpQoE=
-    // example.com.		3600	IN	RRSIG	DNSKEY 8 2 3600 20150914090528 20150824071818 31406 example.com. rZJRBwHhYzCDwkDEXqECHNWezTNj2A683I/yHHqD1j9ytGHGskGEEyJC i5fk70YCm64GqDYKu70kgv7hCFqc4OM3aD88QDe3L4Uv7ZXqouNbjTEO 3BEBI13GetRkK5qLndl30Y/urOBASQFELQUJsvQBR2gJMdQsb6G0mHIW rubY2SxAGa9rQW7yehRQNK4ME37FqINBDuIV9o7kULPhn9Ux1Qx62prd 9nikzamGxFL+9dFDOfnYVw2C/OgGJNIXh5QyKMG4qXmXb6sB/V3P+FE+ +vkt3RToE2xPN5bf1vVIlEJof6LtojrowwnZpiphTXFJF/BJrgiotGt3 Gsd8Cw==
-    // example.com.		3600	IN	DNSKEY	256 3 8 AwEAAcZMEndf6/+kG6Dp7re/grJ9f5CP5bQplBGokyxbM4oPNeBfWMIC +xY+ICgTyJarVB4aPYNMV7znsHM4XwU8hfpZ3ZcmT+69KyGqs+tt2pc/ si30dnUpPo/AMnN7Kul2SgqT9g1bb5O0D/CH2txo6YXr/BbuNHLqAh/x mof1QYkl6GoP
-    // example.com.		3600	IN	DNSKEY	256 3 8 AwEAAeZFCLkW/sztmJmpmZo/udvAyqshiLO34zHzzkVPrhuUBA/xb3wk YeCvMO6iBxCD+/Dk7fWEAT1NR21bDKHySVHE5cre+fqnXI+9NCjkMoBE 193j8G5HscIpWpG1qgkelBhmucfUPv+R4AIhpfjc352eh1q/SniYUGR4 fytlDZVXCLhL
-    // example.com.		3600	IN	DNSKEY	257 3 8 AwEAAbOFAxl+Lkt0UMglZizKEC1AxUu8zlj65KYatR5wBWMrh18TYzK/ ig6Y1t5YTWCO68bynorpNu9fqNFALX7bVl9/gybA0v0EhF+dgXmoUfRX 7ksMGgBvtfa2/Y9a3klXNLqkTszIQ4PEMVCjtryl19Be9/PkFeC9ITjg MRQsQhmB39eyMYnal+f3bUxKk4fq7cuEU0dbRpue4H/N6jPucXWOwiMA kTJhghqgy+o9FfIp+tR/emKao94/wpVXDcPf5B18j7xz2SvTTxiuqCzC MtsxnikZHcoh1j4g+Y1B8zIMIvrEM+pZGhh/Yuf4RwCBgaYCi9hpiMWV vS4WBzx0/lU=
-    // example.com.		3600	IN	RRSIG	SOA 8 2 3600 20150926132522 20150905040848 54108 example.com. q8psdDPaJVo9KPVgMNR2N1by3LMEci+3HyTmN/Xv3DgDFG5MqNlX9Dfj dUBIMbvYmkUUPQ9fIWYA+ldmDHiRBiHIcvvk/LYD8mODWL6RoF+GEsW0 zm43RNBnbE41wtNrch5WU/q1ko2svB98ooqePWWuFzmdyPpidtLCgSCz FCiCiVQ=
-
-    // www
-    let www_name: Name = Name::parse("www.example.com.", None).unwrap();
-
-    // www.example.com.	86400	IN	TXT	"v=spf1 -all"
-    records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["v=spf1 -all".to_string()]))).clone(), 0);
-
-    // www.example.com.	86400	IN	A	93.184.216.34
-    records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone(), 0);
-
-    // www.example.com.	86400	IN	AAAA	2606:2800:220:1:248:1893:25c8:1946
-    records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(), 0);
-
-    // www.example.com.	3600	IN	RRSIG	NSEC 8 3 3600 20150925215757 20150905040848 54108 example.com. ZKIVt1IN3O1FWZPSfrQAH7nHt7RUFDjcbh7NxnEqd/uTGCnZ6SrAEgrY E9GMmBwvRjoucphGtjkYOpPJPe5MlnTHoYCjxL4qmG3LsD2KD0bfPufa ibtlQZRrPglxZ92hBKK3ZiPnPRe7I9yni2UQSQA7XDi7CQySYyo490It AxdXjAo=
-    // www.example.com.	3600	IN	NSEC	example.com. A TXT AAAA RRSIG NSEC
-    // www.example.com.	86400	IN	RRSIG	TXT 8 3 86400 20150914142952 20150824191224 54108 example.com. LvODnPb7NLDZfHPBOrr/qLnOKA670vVYKQSk5Qkz3MPNKDVAFJqsP2Y6 UYcypSJZfcSjfIk2mU9dUiansU2ZL80OZJUsUobqJt5De748ovITYDJ7 afbohQzPg+4E1GIWMkJZ/VQD3B2pmr7J5rPn+vejxSQSoI93AIQaTpCU L5O/Bac=
-    // www.example.com.	86400	IN	RRSIG	AAAA 8 3 86400 20150914082216 20150824191224 54108 example.com. kje4FKE+7d/j4OzWQelcKkePq6DxCRY/5btAiUcZNf+zVNlHK+o57h1r Y76ZviWChQB8Np2TjA1DrXGi/kHr2KKE60H5822mFZ2b9O+sgW4q6o3G kO2E1CQxbYe+nI1Z8lVfjdCNm81zfvYqDjo2/tGqagehxG1V9MBZO6br 4KKdoa4=
-    // www.example.com.	86400	IN	RRSIG	A 8 3 86400 20150915023456 20150824191224 54108 example.com. cWtw0nMvcXcYNnxejB3Le3KBfoPPQZLmbaJ8ybdmzBDefQOm1ZjZZMOP wHEIxzdjRhG9mLt1mpyo1H7OezKTGX+mDtskcECTl/+jB/YSZyvbwRxj e88Lrg4D+D2MiajQn3XSWf+6LQVe1J67gdbKTXezvux0tRxBNHHqWXRk pxCILes=
-
-    return records;
-  }
-
-  pub fn create_secure_example() -> Authority {
-    use chrono::Duration;
-    use openssl::crypto::rsa::RSA;
-    use ::rr::dnssec::{Algorithm, Signer};
-
-    let mut authority: Authority = create_example();
-    let rsa = RSA::generate(2048).unwrap();
-    let signer = Signer::new(Algorithm::RSASHA256, rsa, authority.get_origin().clone(), Duration::weeks(1));
-
-    authority.add_secure_key(signer);
-    authority.secure_zone();
-
-    authority
-  }
-
-  #[test]
-  fn test_search() {
-    let example = create_example();
-    let origin = example.get_origin().clone();
-
-    let mut query: Query = Query::new();
-    query.name(origin.clone());
-
-    let result = example.search(&query, false);
-    if !result.is_empty() {
-      assert_eq!(result.first().unwrap().get_rr_type(), RecordType::A);
-      assert_eq!(result.first().unwrap().get_dns_class(), DNSClass::IN);
-      assert_eq!(result.first().unwrap().get_rdata(), &RData::A(Ipv4Addr::new(93,184,216,34)));
-    } else {
-      panic!("expected a result");  // valid panic, in test
-    }
-  }
-
-  /// this is a litte more interesting b/c it requires a recursive lookup for the origin
-  #[test]
-  fn test_search_www() {
-    let example = create_example();
-    let www_name = Name::parse("www.example.com.", None).unwrap();
-
-    let mut query: Query = Query::new();
-    query.name(www_name.clone());
-
-    let result = example.search(&query, false);
-    if !result.is_empty() {
-      assert_eq!(result.first().unwrap().get_rr_type(), RecordType::A);
-      assert_eq!(result.first().unwrap().get_dns_class(), DNSClass::IN);
-      assert_eq!(result.first().unwrap().get_rdata(), &RData::A(Ipv4Addr::new(93,184,216,34)));
-    } else {
-      panic!("expected a result");  // valid panic, in test
-    }
-  }
-
-  #[test]
-  fn test_authority() {
-    let authority: Authority = create_example();
-
-    assert!(authority.get_soa().is_some());
-    assert_eq!(authority.get_soa().unwrap().get_dns_class(), DNSClass::IN);
-
-    assert!(!authority.lookup(authority.get_origin(), RecordType::NS, false).is_empty());
-
-    let mut lookup: Vec<_> = authority.get_ns(false);
-    lookup.sort();
-
-    assert_eq!(**lookup.first().unwrap(), Record::new().name(authority.get_origin().clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()) ).clone());
-    assert_eq!(**lookup.last().unwrap(), Record::new().name(authority.get_origin().clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("b.iana-servers.net.", None).unwrap()) ).clone());
-
-    assert!(!authority.lookup(authority.get_origin(), RecordType::TXT, false).is_empty());
-
-    let mut lookup: Vec<_> = authority.lookup(authority.get_origin(), RecordType::TXT, false);
-    lookup.sort();
-
-    assert_eq!(**lookup.first().unwrap(), Record::new().name(authority.get_origin().clone()).ttl(60).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["$Id: example.com 4415 2015-08-24 20:12:23Z davids $".to_string()]))).clone());
-
-    assert_eq!(**authority.lookup(authority.get_origin(), RecordType::A, false).first().unwrap(), Record::new().name(authority.get_origin().clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone());
-  }
-
-  #[test]
-  fn test_authorize() {
-    let authority: Authority = create_example();
-
-    let mut message = Message::new();
-    message.id(10).message_type(MessageType::Query).op_code(OpCode::Update);
-
-    assert_eq!(authority.authorize(&message), Err(ResponseCode::Refused));
-
-    // TODO: this will nee to be more complex as additional policies are added
-    // authority.set_allow_update(true);
-    // assert!(authority.authorize(&message).is_ok());
-  }
-
-  #[test]
-  fn test_prerequisites() {
-    let not_zone = Name::new().label("not").label("a").label("domain").label("com");
-    let not_in_zone = Name::new().label("not").label("example").label("com");
-
-    let mut authority: Authority = create_example();
-    authority.set_allow_update(true);
-
-    // first check the initial negatives, ttl = 0, and the zone is the same
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(not_zone.clone()).ttl(0).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::NotZone));
-
-    // *   ANY      ANY      empty    Name is in use
-    assert!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::ANY).rr_type(RecordType::ANY).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(0).dns_class(DNSClass::ANY).rr_type(RecordType::ANY).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::NXDomain));
-
-    // *   ANY      rrset    empty    RRset exists (value independent)
-    assert!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::ANY).rr_type(RecordType::A).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(0).dns_class(DNSClass::ANY).rr_type(RecordType::A).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::NXRRSet));
-
-    // *   NONE     ANY      empty    Name is not in use
-    assert!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(0).dns_class(DNSClass::NONE).rr_type(RecordType::ANY).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::NONE).rr_type(RecordType::ANY).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::YXDomain));
-
-    // *   NONE     rrset    empty    RRset does not exist
-    assert!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(0).dns_class(DNSClass::NONE).rr_type(RecordType::A).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::NONE).rr_type(RecordType::A).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::YXRRSet));
-
-    // *   zone     rrset    rr       RRset exists (value dependent)
-    assert!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::IN).rr_type(RecordType::A).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone()]).is_ok());
-    // wrong class
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::CH).rr_type(RecordType::A).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone()]), Err(ResponseCode::FormErr));
-    // wrong Name
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(not_in_zone.clone()).ttl(0).dns_class(DNSClass::IN).rr_type(RecordType::A).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]), Err(ResponseCode::NXRRSet));
-    // wrong IP
-    assert_eq!(authority.verify_prerequisites(&[Record::new().name(authority.get_origin().clone()).ttl(0).dns_class(DNSClass::IN).rr_type(RecordType::A).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]), Err(ResponseCode::NXRRSet));
-  }
-
-  #[test]
-  fn test_pre_scan() {
-    let up_name = Name::new().label("www").label("example").label("com");
-    let not_zone = Name::new().label("not").label("zone").label("com");
-
-    let authority: Authority = create_example();
-
-    assert_eq!(authority.pre_scan(&[Record::new().name(not_zone.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]), Err(ResponseCode::NotZone));
-
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::ANY).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new()) ).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::AXFR).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new()) ).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::IXFR).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new()) ).clone()]), Err(ResponseCode::FormErr));
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]).is_ok());
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::ANY).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::A).dns_class(DNSClass::ANY).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::AXFR).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::IXFR).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::ANY).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::A).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::NONE).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::ANY).dns_class(DNSClass::NONE).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::AXFR).dns_class(DNSClass::NONE).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::IXFR).dns_class(DNSClass::NONE).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::A).dns_class(DNSClass::NONE).rdata(RData::NULL(NULL::new())).clone()]).is_ok());
-    assert!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(0).rr_type(RecordType::A).dns_class(DNSClass::NONE).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()]).is_ok());
-
-    assert_eq!(authority.pre_scan(&[Record::new().name(up_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::CH).rdata(RData::NULL(NULL::new())).clone()]), Err(ResponseCode::FormErr));
-  }
-
-  #[test]
-  fn test_update() {
-    let new_name = Name::new().label("new").label("example").label("com");
-    let www_name = Name::new().label("www").label("example").label("com");
-    let mut authority: Authority = create_example();
-    let serial = authority.get_serial();
-
-    authority.set_allow_update(true);
-
-    let mut original_vec: Vec<Record> = vec![
-      Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["v=spf1 -all".to_string()]))).clone(),
-      Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone(),
-      Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(),
-    ];
-
-    original_vec.sort();
-
-    {
-      // assert that the correct set of records is there.
-      let mut www_rrset: Vec<&Record> = authority.lookup(&www_name, RecordType::ANY, false);
-      www_rrset.sort();
-
-      assert_eq!(www_rrset, original_vec.iter().collect::<Vec<&Record>>());
-
-      // assert new record doesn't exist
-      assert!(authority.lookup(&new_name, RecordType::ANY, false).is_empty());
-    }
-
-    //
-    //  zone     rrset    rr       Add to an RRset
-    let add_record = &[Record::new().name(new_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()];
-    assert!(authority.update_records(add_record, true).expect("update failed"));
-    assert_eq!(authority.lookup(&new_name, RecordType::ANY, false), add_record.iter().collect::<Vec<&Record>>());
-    assert_eq!(serial + 1, authority.get_serial());
-
-    let add_www_record = &[Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(10,0,0,1))).clone()];
-    assert!(authority.update_records(add_www_record, true).expect("update failed"));
-    assert_eq!(serial + 2, authority.get_serial());
-
-    {
-      let mut www_rrset = authority.lookup(&www_name, RecordType::ANY, false);
-      www_rrset.sort();
-
-      let mut plus_10 = original_vec.clone();
-      plus_10.push(add_www_record[0].clone());
-      plus_10.sort();
-      assert_eq!(www_rrset, plus_10.iter().collect::<Vec<&Record>>());
-    }
-
-    //
-    //  NONE     rrset    rr       Delete an RR from an RRset
-    let del_record = &[Record::new().name(new_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::NONE).rdata(RData::A(Ipv4Addr::new(93,184,216,24))).clone()];
-    assert!(authority.update_records(del_record, true).expect("update failed"));
-    assert_eq!(serial + 3, authority.get_serial());
-    {
-      println!("after delete of specific record: {:?}", authority.lookup(&new_name, RecordType::ANY, false));
-      assert!(authority.lookup(&new_name, RecordType::ANY, false).is_empty());
-    }
-
-    // remove one from www
-    let del_record = &[Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::NONE).rdata(RData::A(Ipv4Addr::new(10,0,0,1))).clone()];
-    assert!(authority.update_records(del_record, true).expect("update failed"));
-    assert_eq!(serial + 4, authority.get_serial());
-    {
-      let mut www_rrset = authority.lookup(&www_name, RecordType::ANY, false);
-      www_rrset.sort();
-
-      assert_eq!(www_rrset, original_vec.iter().collect::<Vec<&Record>>());
-    }
-
-    //
-    //  ANY      rrset    empty    Delete an RRset
-    let del_record = &[Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()];
-    assert!(authority.update_records(del_record, true).expect("update failed"));
-    assert_eq!(serial + 5, authority.get_serial());
-    let mut removed_a_vec: Vec<_> = vec![
-      Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["v=spf1 -all".to_string()]))).clone(),
-      Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(),
-    ];
-    removed_a_vec.sort();
-
-    {
-      let mut www_rrset = authority.lookup(&www_name, RecordType::ANY, false);
-      www_rrset.sort();
-
-      assert_eq!(www_rrset, removed_a_vec.iter().collect::<Vec<&Record>>());
-    }
-
-    //
-    //  ANY      ANY      empty    Delete all RRsets from a name
-    println!("deleting all records");
-    let del_record = &[Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::ANY).dns_class(DNSClass::ANY).rdata(RData::NULL(NULL::new())).clone()];
-    assert!(authority.update_records(del_record, true).expect("update failed"));
-    assert!(authority.lookup(&www_name, RecordType::ANY, false).is_empty());
-    assert_eq!(serial + 6, authority.get_serial());
-  }
-
-  #[test]
-  fn test_zone_signing() {
-    use ::rr::{RData};
-
-    let authority: Authority = create_secure_example();
-
-    let results = authority.lookup(&authority.get_origin(), RecordType::AXFR, true);
-
-    assert!(results.iter().any(|r| r.get_rr_type() == RecordType::DNSKEY), "must contain a DNSKEY");
-
-    for record in results.iter() {
-      if record.get_rr_type() == RecordType::RRSIG { continue }
-      if record.get_rr_type() == RecordType::DNSKEY { continue }
-
-      // validate all records have associated RRSIGs after signing
-      assert!(results.iter().any(|r| r.get_rr_type() == RecordType::RRSIG &&
-                                     r.get_name() == record.get_name() &&
-                                     if let &RData::SIG(ref rrsig) = r.get_rdata() {
-                                       rrsig.get_type_covered() == record.get_rr_type()
-                                     } else {
-                                       false
-                                     } ), "record type not covered: {:?}", record);
-    }
-  }
-
-  #[test]
-  fn test_get_nsec() {
-    let name = Name::new().label("zzz").label("example").label("com");
-    let authority: Authority = create_secure_example();
-
-    let results = authority.get_nsec_records(&name, true);
-
-    for record in results.iter() {
-      assert!(record.get_name() < &name);
-    }
-  }
-
-  #[test]
-  fn test_journal() {
-    use std::net::Ipv4Addr;
-    use rusqlite::Connection;
-    use ::authority::Journal;
-
-    // test that this message can be inserted
-    let conn = Connection::open_in_memory().expect("could not create in memory DB");
-    let mut journal = Journal::new(conn).unwrap();
-    journal.schema_up().unwrap();
-
-    let mut authority = create_example();
-    authority.journal(journal);
-    authority.persist_to_journal().unwrap();
-
-    let new_name = Name::new().label("new").label("example").label("com");
-    let delete_name = Name::new().label("www").label("example").label("com");
-    let new_record = Record::new().name(new_name.clone()).rdata(RData::A(Ipv4Addr::new(10,11,12,13))).clone();
-    let delete_record = Record::new().name(delete_name.clone()).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).dns_class(DNSClass::NONE).clone();
-    authority.update_records(&[new_record.clone(), delete_record], true).unwrap();
-
-    // assert that the correct set of records is there.
-    let new_rrset: Vec<&Record> = authority.lookup(&new_name, RecordType::A, false);
-    assert!(new_rrset.iter().all(|r| *r == &new_record));
-
-    let delete_rrset: Vec<&Record> = authority.lookup(&delete_name, RecordType::A, false);
-    assert!(delete_rrset.is_empty());
-
-    // that record should have been recorded... let's reload the journal and see if we get it.
-    let mut recovered_authority = Authority::new(authority.get_origin().clone(),
-                                                 BTreeMap::new(),
-                                                 ZoneType::Master,
-                                                 false);
-    recovered_authority.recover_with_journal(authority.get_journal().expect("journal not Some")).expect("recovery");
-
-    // assert that the correct set of records is there.
-    let new_rrset: Vec<&Record> = recovered_authority.lookup(&new_name, RecordType::A, false);
-    assert!(new_rrset.iter().all(|r| *r == &new_record));
-
-    let delete_rrset: Vec<&Record> = authority.lookup(&delete_name, RecordType::A, false);
-    assert!(delete_rrset.is_empty());
-  }
-
-  #[test]
-  fn test_recovery() {
-    use rusqlite::Connection;
-    use ::authority::Journal;
-
-    // test that this message can be inserted
-    let conn = Connection::open_in_memory().expect("could not create in memory DB");
-    let mut journal = Journal::new(conn).unwrap();
-    journal.schema_up().unwrap();
-
-    let mut authority = create_example();
-    authority.journal(journal);
-    authority.persist_to_journal().unwrap();
-
-    let journal = authority.get_journal().unwrap();
-    let mut recovered_authority = Authority::new(authority.get_origin().clone(),
-                                                 BTreeMap::new(),
-                                                 ZoneType::Master,
-                                                 false);
-
-    recovered_authority.recover_with_journal(journal).expect("recovery");
-
-    assert_eq!(recovered_authority.get_records().len(), authority.get_records().len());
-    assert_eq!(recovered_authority.get_soa(), authority.get_soa());
-    assert!(recovered_authority.get_records().iter().all(|(rr_key, rr_set)| {
-      let other_rr_set = authority.get_records().get(rr_key).expect(&format!("key doesn't exist: {:?}", rr_key));
-      rr_set.iter().zip(other_rr_set.iter()).all(|(record, other_record)| {
-        record.get_ttl() == other_record.get_ttl() &&
-        record.get_rdata() == other_record.get_rdata()
-      })
-    }));
-
-    assert!(authority.get_records().iter().all(|(rr_key, rr_set)| {
-      let other_rr_set = recovered_authority.get_records().get(rr_key).expect(&format!("key doesn't exist: {:?}", rr_key));
-      rr_set.iter().zip(other_rr_set.iter()).all(|(record, other_record)| {
-        record.get_ttl() == other_record.get_ttl() &&
-        record.get_rdata() == other_record.get_rdata()
-      })
-    }));
-  }
+pub fn create_example() -> Authority {
+  use std::net::*;
+  use trust_dns::rr::rdata::*;
+
+  let origin: Name = Name::parse("example.com.", None,).unwrap();
+  let mut records: Authority = Authority::new(origin.clone(), BTreeMap::new(), ZoneType::Master, false);
+  // example.com.		3600	IN	SOA	sns.dns.icann.org. noc.dns.icann.org. 2015082403 7200 3600 1209600 3600
+  records.upsert(Record::new().name(origin.clone()).ttl(3600).rr_type(RecordType::SOA).dns_class(DNSClass::IN).rdata(RData::SOA(SOA::new(Name::parse("sns.dns.icann.org.", None).unwrap(), Name::parse("noc.dns.icann.org.", None).unwrap(), 2015082403, 7200, 3600, 1209600, 3600 ))).clone(), 0);
+
+  records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()) ).clone(), 0);
+  records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::NS).dns_class(DNSClass::IN).rdata(RData::NS(Name::parse("b.iana-servers.net.", None).unwrap()) ).clone(), 0);
+
+  // example.com.		60	IN	TXT	"v=spf1 -all"
+  //records.upsert(origin.clone(), Record::new().name(origin.clone()).ttl(60).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT{ txt_data: vec!["v=spf1 -all".to_string()] }).clone());
+  // example.com.		60	IN	TXT	"$Id: example.com 4415 2015-08-24 20:12:23Z davids $"
+  records.upsert(Record::new().name(origin.clone()).ttl(60).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["$Id: example.com 4415 2015-08-24 20:12:23Z davids $".to_string()]))).clone(), 0);
+
+  // example.com.		86400	IN	A	93.184.216.34
+  records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone(), 0);
+
+  // example.com.		86400	IN	AAAA	2606:2800:220:1:248:1893:25c8:1946
+  records.upsert(Record::new().name(origin.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(), 0);
+
+  // TODO support these later...
+
+  // example.com.		3600	IN	RRSIG	NSEC 8 2 3600 20150926015219 20150905040848 54108 example.com. d0AXd6QRITqLeiYbQUlJ5O0Og9tSjk7IlxQr9aJO+r+rc1g0dW9i9OCc XXQxdC1/zyubecjD6kSs3vwxzzEEupivaKHKtNPXdnDZ5UUiaIC1VU9l 9h/ik+AR4rCTY6dYPCI6lafD/TlqQLbpEnb34ywkRpl5G3pasPrwEY7b nrAndEY=
+  // example.com.		3600	IN	NSEC	www.example.com. A NS SOA TXT AAAA RRSIG NSEC DNSKEY
+  // example.com.		86400	IN	RRSIG	NS 8 2 86400 20150915033148 20150824191224 54108 example.com. O2TCB5/v/b1XGlTQEj0/oGKp7dTueQ7zRmCtADDEDWrzLdWrKcmDGF37 mgKejcAlSYVhWLxyLlet7KqJhLu+oQcDTNf/BT3vNX/Ivx3sKhUUMpfi 8Mn5zhRqM9gbzZVCS/toJIYqOBqvAkS7UpkmpLzl0Zt2h4j0Gp/8GwRb ZU67l6M=
+  // example.com.		86400	IN	RRSIG	AAAA 8 2 86400 20150914212400 20150824191224 54108 example.com. AHd2BDNjtg4jPRQwyT4FHtlVTZDZ6IIusYVGCzWfnt5SZOoizyXnJhqX 44MeVTqi1/2cskpKvRkK3bkYnVUcjZiFgSaa9xJHmXrslaTr5mOmXt9s 6k95N1daYKhDKKcr0M4TXLUgdnBr+/pMFiLsyOoDb8GJDT8Llmpk52Ie ysJX8BY=
+  // example.com.		86400	IN	RRSIG	A 8 2 86400 20150914083326 20150824191224 54108 example.com. La1p2R7GPMrXEm3kcznSJ70sOspmfSDsgOZ74GlzgaFfMRveA20IDUnZ /HI9M95/tBWbHdHBtm9aCK+4n7EluhNPTAT1+88V6xK7Lc7pcBfBXIHg DAdUoj26VIh7NRml/0QR0dFu4PriA/wLNe+d1Q961qf0JZP80TU4IMBC X/W6Ijk=
+  // example.com.		60	IN	RRSIG	TXT 8 2 60 20150914201612 20150824191224 54108 example.com. Be/bPvaVVK/o66QOHJZMFBDCQVhP44jptS9sZe8Vpfmzd72/v+1gwn1z u2+xisePSpAMtDZsFJgqsCjpbLFvmhNdh8ktlq/kuCME5hZs7qY7DZIB VwkSTsJPIq8qhX22clfIbqzaypuIX9ajWr+5i0nGQLNekMB07t4/GCoJ q5QpQoE=
+  // example.com.		3600	IN	RRSIG	DNSKEY 8 2 3600 20150914090528 20150824071818 31406 example.com. rZJRBwHhYzCDwkDEXqECHNWezTNj2A683I/yHHqD1j9ytGHGskGEEyJC i5fk70YCm64GqDYKu70kgv7hCFqc4OM3aD88QDe3L4Uv7ZXqouNbjTEO 3BEBI13GetRkK5qLndl30Y/urOBASQFELQUJsvQBR2gJMdQsb6G0mHIW rubY2SxAGa9rQW7yehRQNK4ME37FqINBDuIV9o7kULPhn9Ux1Qx62prd 9nikzamGxFL+9dFDOfnYVw2C/OgGJNIXh5QyKMG4qXmXb6sB/V3P+FE+ +vkt3RToE2xPN5bf1vVIlEJof6LtojrowwnZpiphTXFJF/BJrgiotGt3 Gsd8Cw==
+  // example.com.		3600	IN	DNSKEY	256 3 8 AwEAAcZMEndf6/+kG6Dp7re/grJ9f5CP5bQplBGokyxbM4oPNeBfWMIC +xY+ICgTyJarVB4aPYNMV7znsHM4XwU8hfpZ3ZcmT+69KyGqs+tt2pc/ si30dnUpPo/AMnN7Kul2SgqT9g1bb5O0D/CH2txo6YXr/BbuNHLqAh/x mof1QYkl6GoP
+  // example.com.		3600	IN	DNSKEY	256 3 8 AwEAAeZFCLkW/sztmJmpmZo/udvAyqshiLO34zHzzkVPrhuUBA/xb3wk YeCvMO6iBxCD+/Dk7fWEAT1NR21bDKHySVHE5cre+fqnXI+9NCjkMoBE 193j8G5HscIpWpG1qgkelBhmucfUPv+R4AIhpfjc352eh1q/SniYUGR4 fytlDZVXCLhL
+  // example.com.		3600	IN	DNSKEY	257 3 8 AwEAAbOFAxl+Lkt0UMglZizKEC1AxUu8zlj65KYatR5wBWMrh18TYzK/ ig6Y1t5YTWCO68bynorpNu9fqNFALX7bVl9/gybA0v0EhF+dgXmoUfRX 7ksMGgBvtfa2/Y9a3klXNLqkTszIQ4PEMVCjtryl19Be9/PkFeC9ITjg MRQsQhmB39eyMYnal+f3bUxKk4fq7cuEU0dbRpue4H/N6jPucXWOwiMA kTJhghqgy+o9FfIp+tR/emKao94/wpVXDcPf5B18j7xz2SvTTxiuqCzC MtsxnikZHcoh1j4g+Y1B8zIMIvrEM+pZGhh/Yuf4RwCBgaYCi9hpiMWV vS4WBzx0/lU=
+  // example.com.		3600	IN	RRSIG	SOA 8 2 3600 20150926132522 20150905040848 54108 example.com. q8psdDPaJVo9KPVgMNR2N1by3LMEci+3HyTmN/Xv3DgDFG5MqNlX9Dfj dUBIMbvYmkUUPQ9fIWYA+ldmDHiRBiHIcvvk/LYD8mODWL6RoF+GEsW0 zm43RNBnbE41wtNrch5WU/q1ko2svB98ooqePWWuFzmdyPpidtLCgSCz FCiCiVQ=
+
+  // www
+  let www_name: Name = Name::parse("www.example.com.", None).unwrap();
+
+  // www.example.com.	86400	IN	TXT	"v=spf1 -all"
+  records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::TXT).dns_class(DNSClass::IN).rdata(RData::TXT(TXT::new(vec!["v=spf1 -all".to_string()]))).clone(), 0);
+
+  // www.example.com.	86400	IN	A	93.184.216.34
+  records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::A).dns_class(DNSClass::IN).rdata(RData::A(Ipv4Addr::new(93,184,216,34))).clone(), 0);
+
+  // www.example.com.	86400	IN	AAAA	2606:2800:220:1:248:1893:25c8:1946
+  records.upsert(Record::new().name(www_name.clone()).ttl(86400).rr_type(RecordType::AAAA).dns_class(DNSClass::IN).rdata(RData::AAAA(Ipv6Addr::new(0x2606,0x2800,0x220,0x1,0x248,0x1893,0x25c8,0x1946))).clone(), 0);
+
+  // www.example.com.	3600	IN	RRSIG	NSEC 8 3 3600 20150925215757 20150905040848 54108 example.com. ZKIVt1IN3O1FWZPSfrQAH7nHt7RUFDjcbh7NxnEqd/uTGCnZ6SrAEgrY E9GMmBwvRjoucphGtjkYOpPJPe5MlnTHoYCjxL4qmG3LsD2KD0bfPufa ibtlQZRrPglxZ92hBKK3ZiPnPRe7I9yni2UQSQA7XDi7CQySYyo490It AxdXjAo=
+  // www.example.com.	3600	IN	NSEC	example.com. A TXT AAAA RRSIG NSEC
+  // www.example.com.	86400	IN	RRSIG	TXT 8 3 86400 20150914142952 20150824191224 54108 example.com. LvODnPb7NLDZfHPBOrr/qLnOKA670vVYKQSk5Qkz3MPNKDVAFJqsP2Y6 UYcypSJZfcSjfIk2mU9dUiansU2ZL80OZJUsUobqJt5De748ovITYDJ7 afbohQzPg+4E1GIWMkJZ/VQD3B2pmr7J5rPn+vejxSQSoI93AIQaTpCU L5O/Bac=
+  // www.example.com.	86400	IN	RRSIG	AAAA 8 3 86400 20150914082216 20150824191224 54108 example.com. kje4FKE+7d/j4OzWQelcKkePq6DxCRY/5btAiUcZNf+zVNlHK+o57h1r Y76ZviWChQB8Np2TjA1DrXGi/kHr2KKE60H5822mFZ2b9O+sgW4q6o3G kO2E1CQxbYe+nI1Z8lVfjdCNm81zfvYqDjo2/tGqagehxG1V9MBZO6br 4KKdoa4=
+  // www.example.com.	86400	IN	RRSIG	A 8 3 86400 20150915023456 20150824191224 54108 example.com. cWtw0nMvcXcYNnxejB3Le3KBfoPPQZLmbaJ8ybdmzBDefQOm1ZjZZMOP wHEIxzdjRhG9mLt1mpyo1H7OezKTGX+mDtskcECTl/+jB/YSZyvbwRxj e88Lrg4D+D2MiajQn3XSWf+6LQVe1J67gdbKTXezvux0tRxBNHHqWXRk pxCILes=
+
+  return records;
+}
+
+pub fn create_secure_example() -> Authority {
+  use chrono::Duration;
+  use openssl::crypto::rsa::RSA;
+  use trust_dns::rr::dnssec::*;
+
+  let mut authority: Authority = create_example();
+  let rsa = RSA::generate(2048).unwrap();
+  let signer = Signer::new(Algorithm::RSASHA256, rsa, authority.get_origin().clone(), Duration::weeks(1));
+
+  authority.add_secure_key(signer);
+  authority.secure_zone();
+
+  authority
 }
