+++ server/src/authority/persistence.rs
@@ -12,12 +12,12 @@ use rusqlite;
 use rusqlite::Connection;
 use rusqlite::SqliteError;
 
-use ::error::PersistenceErrorKind;
-use ::error::PersistenceResult;
-use ::rr::Record;
-use ::serialize::binary::{BinDecoder, BinEncoder, BinSerializable};
+use trust_dns::rr::Record;
+use trust_dns::serialize::binary::{BinDecoder, BinEncoder, BinSerializable};
 
-const CURRENT_VERSION: i64 = 1;
+use ::error::{PersistenceErrorKind, PersistenceResult};
+
+pub const CURRENT_VERSION: i64 = 1;
 
 /// The Journal is the audit log of all changes to a zone after initial creation.
 pub struct Journal {
@@ -40,6 +40,10 @@ impl Journal {
       },
       Err(err) => Err(err)
     }
+  }
+
+  pub fn get_conn(&self) -> &Connection {
+    &self.conn
   }
 
   /// gets the current schema version of the journal
@@ -140,7 +144,7 @@ impl Journal {
   /// # Arguments
   ///
   /// * `conn` - db connection to use
-  fn select_schema_version(conn: &Connection) -> PersistenceResult<i64> {
+  pub fn select_schema_version(conn: &Connection) -> PersistenceResult<i64> {
     // first see if our schema is there
     let mut stmt = try!(conn.prepare("SELECT name
                                         FROM sqlite_master
@@ -257,88 +261,4 @@ impl<'j> Iterator for JournalIter<'j> {
       }
     }
   }
-}
-
-#[test]
-fn test_new_journal() {
-  let conn = Connection::open_in_memory().expect("could not create in memory DB");
-  assert_eq!(Journal::new(conn).expect("new Journal").get_schema_version(), -1);
-}
-
-#[test]
-fn test_init_journal() {
-  let conn = Connection::open_in_memory().expect("could not create in memory DB");
-  let mut journal = Journal::new(conn).unwrap();
-  let version = journal.schema_up().unwrap();
-  assert_eq!(version, CURRENT_VERSION);
-  assert_eq!(Journal::select_schema_version(&journal.conn).unwrap(), CURRENT_VERSION);
-}
-
-#[cfg(test)]
-fn create_test_journal() -> (Record, Journal) {
-  use std::net::Ipv4Addr;
-  use std::str::FromStr;
-
-  use ::rr::{Name, RData, Record, RecordType};
-
-  let www = Name::with_labels(vec!["www".to_string(),"example".to_string(), "com".to_string()]);
-
-  let mut record = Record::new();
-  record.name(www);
-  record.rr_type(RecordType::A);
-  record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap()));
-
-  // test that this message can be inserted
-  let conn = Connection::open_in_memory().expect("could not create in memory DB");
-  let mut journal = Journal::new(conn).unwrap();
-  journal.schema_up().unwrap();
-
-  // insert the message
-  journal.insert_record(0, &record).unwrap();
-
-  // insert another...
-  record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap()));
-  journal.insert_record(0, &record).unwrap();
-
-  (record, journal)
-}
-
-#[test]
-fn test_insert_and_select_record() {
-  use std::net::Ipv4Addr;
-  use std::str::FromStr;
-
-  use rr::RData;
-
-  let (mut record, journal) = create_test_journal();
-
-  // select the record
-  let (row_id, journal_record) = journal.select_record(0).expect("persistence error").expect("none");
-  record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap()));
-  assert_eq!(journal_record, record);
-
-  // test another
-  let (row_id, journal_record) = journal.select_record(row_id + 1).expect("persistence error").expect("none");
-  record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap()));
-  assert_eq!(journal_record, record);
-
-  // check that we get nothing for id over row_id
-  let option_none = journal.select_record(row_id + 1).expect("persistence error");
-  assert!(option_none.is_none());
-}
-
-#[test]
-fn test_iterator() {
-  use std::net::Ipv4Addr;
-  use std::str::FromStr;
-
-  use rr::RData;
-
-  let (mut record, journal) = create_test_journal();
-
-  let mut iter = journal.iter();
-
-  assert_eq!(record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap())), &iter.next().unwrap());
-  assert_eq!(record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap())), &iter.next().unwrap());
-  assert_eq!(None, iter.next());
 }
